:sparkles: Added OtherUserAccountPage
- Created structure for setting user privileges
- Added QMLUserModel
- Corrected bug with ErrorDialog

+++ src/rrcore/sqlmanager/usersqlmanager.cpp
@@ -9,6 +9,8 @@
 #include "database/databaseexception.h"
 #include "database/databaseutils.h"
 #include "config/config.h"
+#include "singletons/userprofile.h"
+#include "json/userprivilegecenter.h"
 
 UserSqlManager::UserSqlManager(const QString &connectionName)
     : AbstractSqlManager(connectionName)
@@ -29,6 +31,14 @@ QueryResult UserSqlManager::execute(const QueryRequest &request)
             signUpRootUser(request);
         else if (request.command() == "remove_user")
             removeUser(request);
+        else if (request.command() == "view_users")
+            viewUsers(request, result);
+        else if (request.command() == "view_user_privileges")
+            viewUserPrivileges(request, result);
+        else if (request.command() == "add_user")
+            addUser(request);
+        else if (request.command() == "activate_user")
+            activateUser(request);
         else
             throw DatabaseException(DatabaseException::RRErrorCode::CommandNotFound, QString("Command not found: %1").arg(request.command()));
 
@@ -43,25 +53,47 @@ QueryResult UserSqlManager::execute(const QueryRequest &request)
     return result;
 }
 
-bool UserSqlManager::storeProfile(QueryResult &result)
+bool UserSqlManager::storeProfile(QueryResult &result, const QString &userName)
 {
     QSqlDatabase connection = QSqlDatabase::database(connectionName());
-    if (connection.isOpen()) {
-        QSqlQuery q(connection);
-        q.prepare("SELECT id, user FROM user");
-        if (!q.exec())
-            return false;
 
-        if (q.first())
-            result.setOutcome(recordToMap(q.record()));
-
-        return true;
+    if (!connection.isOpen()) {
+        qWarning() << "Database connection is closed.";
+        throw DatabaseException(DatabaseException::RRErrorCode::Unknown,
+                                QStringLiteral("Database connection is closed."));
     }
 
-    return false;
+    try {
+        const QList<QSqlRecord> &records(callProcedure("GetUserDetails", {
+                                                           ProcedureArgument {
+                                                               ProcedureArgument::Type::In,
+                                                               "user_name",
+                                                               userName
+                                                           }
+                                                       }));
+
+        QVariantMap userDetails;
+        if (!records.isEmpty())
+            userDetails = recordToMap(records.first());
+        else
+            throw DatabaseException(DatabaseException::RRErrorCode::Unknown,
+                                    QStringLiteral("User does not exist."));
+
+        result.setOutcome(QVariantMap {
+                              { "user_name", userName },
+                              { "user_id", userDetails.value("user_id").toInt() },
+                              { "user_privileges", userDetails.value("user_privileges") },
+                              { "record_count", 1 }
+                          });
+    } catch (DatabaseException &e) {
+        qWarning() << "Exception thrown while storing profile:" << e.code() << e.message() << e.userMessage();
+        throw;
+    }
+
+    return true;
 }
 
-void UserSqlManager::signInUser(const QueryRequest &request, QueryResult result)
+void UserSqlManager::signInUser(const QueryRequest &request, QueryResult &result)
 {
     QSqlDatabase connection;
     const QString &userName = request.params().value("user_name").toString();
@@ -79,7 +111,7 @@ void UserSqlManager::signInUser(const QueryRequest &request, QueryResult result)
     connection.setPassword(password);
     connection.setConnectOptions("MYSQL_OPT_RECONNECT = 1");
 
-    if (!connection.open() || !storeProfile(result))
+    if (!connection.open() || !storeProfile(result, userName))
         throw DatabaseException(DatabaseException::RRErrorCode::SignInFailure, connection.lastError().text(),
                                 QString("Failed to sign in as '%1'.").arg(userName));
 }
@@ -215,6 +247,10 @@ void UserSqlManager::removeUser(const QueryRequest &request)
         throw DatabaseException(DatabaseException::RRErrorCode::RemoveUserFailure,
                                 connection.lastError().text(), QString("Connection is closed."));
 
+    if (UserProfile::instance().userId() > 1)
+        throw DatabaseException(DatabaseException::RRErrorCode::InsufficientUserPrivileges,
+                                "You must be a super user to perform this action.");
+
     QSqlQuery q(connection);
     q.prepare("DROP USER :user_name@'localhost'");
     q.bindValue(":user_name", userName);
@@ -223,6 +259,179 @@ void UserSqlManager::removeUser(const QueryRequest &request)
         throw DatabaseException(DatabaseException::RRErrorCode::RemoveUserFailure,
                                 q.lastError().text(),
                                 QString("Failed to drop user '%1'.").arg(userName));
+}
+
+void UserSqlManager::viewUsers(const QueryRequest &request, QueryResult &result)
+{
+    QSqlDatabase connection = QSqlDatabase::database(connectionName());
+    const QVariantMap &params = request.params();
+
+    QSqlQuery q(connection);
+
+    try {
+        const QList<QSqlRecord> &records(callProcedure("ViewUsers", {
+                                                           ProcedureArgument {
+                                                               ProcedureArgument::Type::In,
+                                                               "archived",
+                                                               params.value("archived", false)
+                                                           }
+                                                       }));
+
+        QVariantList users;
+        for (const QSqlRecord &record : records) {
+            users.append(recordToMap(record));
+        }
+
+        result.setOutcome(QVariantMap {
+                              { "users", users },
+                              { "record_count", users.count() }
+                          });
+    } catch (DatabaseException &) {
+        if (!DatabaseUtils::rollbackTransaction(q))
+            qCritical("Failed to rollback failed transaction! %s", q.lastError().text().toStdString().c_str());
+
+        throw;
+    }
+}
+
+void UserSqlManager::viewUserPrivileges(const QueryRequest &request, QueryResult &result)
+{
+    QSqlDatabase connection = QSqlDatabase::database(connectionName());
+    const QVariantMap &params = request.params();
+
+    try {
+        const QList<QSqlRecord> &records(callProcedure("ViewUserPrivileges", {
+                                                           ProcedureArgument {
+                                                               ProcedureArgument::Type::In,
+                                                               "user_id",
+                                                               params.value("user_id", QVariant::Int)
+                                                           }
+                                                       }));
+
+        QVariantMap userPrivileges;
+        for (const QSqlRecord &record : records) {
+            userPrivileges = recordToMap(record);
+        }
+
+        if (userPrivileges.isEmpty()) {
+            UserPrivilegeCenter userPrivilegeCenter;
+            userPrivileges = userPrivilegeCenter.getPrivileges().toMap();
+        }
+
+        result.setOutcome(QVariantMap {
+                              { "user_privileges", userPrivileges },
+                              { "record_count", userPrivileges.count() }
+                          });
+
+    } catch (DatabaseException &) {
+        throw;
+    }
+}
+
+void UserSqlManager::addUser(const QueryRequest &request)
+{
+    QSqlDatabase connection = QSqlDatabase::database(connectionName());
+    const QVariantMap &params = request.params();
+
+    QSqlQuery q(connection);
+
+    try {
+        if (!DatabaseUtils::beginTransaction(q))
+            throw DatabaseException(DatabaseException::RRErrorCode::BeginTransactionFailed, q.lastError().text(), "Failed to start transation.");
+
+        callProcedure("AddSqlUser", {
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "user_name",
+                              params.value("user_name")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "password",
+                              params.value("password")
+                          }
+                      });
+
+        callProcedure("AddRRUser", {
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "first_name",
+                              params.value("first_name")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "last_name",
+                              params.value("last_name")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "user_name",
+                              params.value("user_name")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "photo",
+                              DatabaseUtils::imageToByteArray(params.value("image_url").toString())
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "phone_number",
+                              params.value("phone_number")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "email_address",
+                              params.value("email_address")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "user_id",
+                              UserProfile::instance().userId()
+                          }
+                      });
+
+        if (!DatabaseUtils::commitTransaction(q))
+            throw DatabaseException(DatabaseException::RRErrorCode::CommitTransationFailed, q.lastError().text(), "Failed to commit transation.");
+    } catch (DatabaseException &) {
+        if (!DatabaseUtils::rollbackTransaction(q))
+            qCritical("Failed to rollback failed transaction! %s", q.lastError().text().toStdString().c_str());
+
+        throw;
+    }
+}
+
+void UserSqlManager::activateUser(const QueryRequest &request)
+{
+    QSqlDatabase connection = QSqlDatabase::database(connectionName());
+    const QVariantMap &params = request.params();
+
+    QSqlQuery q(connection);
+
+    try {
+        if (!DatabaseUtils::beginTransaction(q))
+            throw DatabaseException(DatabaseException::RRErrorCode::BeginTransactionFailed, q.lastError().text(), "Failed to start transation.");
+
+        callProcedure("ActivateUser", {
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "active",
+                              params.value("active")
+                          },
+                          ProcedureArgument {
+                              ProcedureArgument::Type::In,
+                              "user_id",
+                              UserProfile::instance().userId()
+                          }
+                      });
+
+        if (!DatabaseUtils::commitTransaction(q))
+            throw DatabaseException(DatabaseException::RRErrorCode::CommitTransationFailed, q.lastError().text(), "Failed to commit transation.");
+    } catch (DatabaseException &) {
+        if (!DatabaseUtils::rollbackTransaction(q))
+            qCritical("Failed to rollback failed transaction! %s", q.lastError().text().toStdString().c_str());
+
+        throw;
+    }
 }
 
 void UserSqlManager::grantPrivilege(const QString &privilege, const QString &userName, QSqlQuery &q)
@@ -240,7 +449,7 @@ void UserSqlManager::createRRUser(const QString &userName, QSqlQuery &q)
     const QDateTime currentDateTime = QDateTime::currentDateTime();
 
     q.prepare(QString("INSERT INTO %1.user (user, photo, phone_number, email_address, active, pending, created, last_edited) "
-              "VALUES (:user, :photo, :phone_number, :email_address, :active, :pending, :created, :last_edited)")
+                      "VALUES (:user, :photo, :phone_number, :email_address, :active, :pending, :created, :last_edited)")
               .arg(Config::instance().databaseName()));
     q.bindValue(":user", userName);
     q.bindValue(":photo", QVariant(QVariant::ByteArray));
@@ -254,4 +463,10 @@ void UserSqlManager::createRRUser(const QString &userName, QSqlQuery &q)
     if (!q.exec())
         throw DatabaseException(DatabaseException::RRErrorCode::SignUpFailure, q.lastError().text(),
                                 QString("Failed to create RR user '%1'.").arg(userName));
+}
+
+QString UserSqlManager::determineRank(const QVariantList &privileges)
+{
+    Q_UNUSED(privileges)
+    return QString();
 }
